xmlparser/libxml: abort if libxml had an error
We cannot do this the intuitive way (i.e. providing a structured
error callback) because, when we call on to the libxml callbacks
any errors within will get reported to us but with the libxml
context pointer, rather than ours. Instead, simply peek inside
the context on return from the libxml handler to see if there
was an error we care about.