bindings/xml/libxml_xmlparser.c: update for libxml2 >= 2.12.0
Version 2.12.0 of libxml2 changes a few functions to return (const
xmlError *) where previously they returned only (xmlError *).
Compilers generally are not happy with this. For example,

  bindings/xml/libxml_xmlparser.c: In function 'xml_parser_start_document':
  bindings/xml/libxml_xmlparser.c:327:16: error: assignment discards 'const'
  qualifier from pointer target type [-Werror=discarded-qualifiers]
  327 |         xmlerr = xmlCtxtGetLastError(parser->xml_ctx);

This commit adds a few #ifdefs to handle both versions of the API
cleanly. It's probably not the sexiest fix, but it's simple and gets
the job done.