node: Fix dom_node_normalize (via _dom_merge_adjacent_text)
Normalizing is meant to turn a sequence of text nodes into one
text node representing the sequence; unfortunately due to the
way that _dom_merge_adjacent_text had been written this actually
resulted in an exponential repetition of the text content.

This is fixed by using the underlying characterdata objects instead.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>